+++ app/src/main/java/woowacourse/paint/model/ColorUiModel.kt
@@ -1,3 +1,5 @@
-package woowacourse.paint
+package woowacourse.paint.model
+
+import woowacourse.paint.canvas.CustomColor
 
 data class ColorUiModel(val color: CustomColor, val isPicked: Boolean = false)
